fix post config script

+++ install_post_config.sh
@@ -1,9 +1,11 @@
 echo "add wget-ssl to executable path to fix helloworld"
-ln -s /usr/xxx/wget-ssl /usr/bin/wget-ssl
+if [ ! -f /usr/bin/wget-ssl ]; then
+  ln -s /usr/xxx/wget-ssl /usr/bin/wget-ssl
+fi
 
 echo "replace sofware source to tsinghua"
-sed -i "s/downloads.openwrt.org/mirrors.tuna.tsinghua.edu.cn\/openwrt/g" feeds.txt
-sed -i "/custom/d" feeds.txt
+sed -i "s/downloads.openwrt.org/mirrors.tuna.tsinghua.edu.cn\/openwrt/g" /etc/opkg/distfeeds.conf
+sed -i "/custom/d" /etc/opkg/distfeeds.conf
 
 echo "install some utilities tools"
 opkg update
@@ -13,5 +15,8 @@ opkg install vim-fuller
 opkg upgrade luci-theme-bootstrap
 
 echo "add luci option to refresh status per second"
-LINE_TO_INSERT=`awk "/main/{print NR}" /etc/config/luci`
-sed -i "${LINE_TO_INSERT}i\ \ \ \ \ \ \ \ option\ pollinterval\ '1'" /etc/config/luci
+grep pollinterval /etc/config/luci
+if [ $? == 1 ]; then
+  LINE_TO_INSERT=`awk "/main/{print NR}" /etc/config/luci`
+  sed -i "${LINE_TO_INSERT} a\ \ \ \ \ \ \ \ option\ pollinterval\ '1'" /etc/config/luci
+fi
